Static-cast cleanup phase 1: drop redundant "as TYPE" casts on numeric-literal statics
Same pattern as the const-cast cleanup phase (cedbbcc), applied to
`static`/`static mut` declarations instead of `const`: c2rust emits
`static mut NAME: TYPE = LITERAL as TYPE;` even though a bare integer
literal already infers TYPE from the static's own declared type.
Mechanically strips `LITERAL as TYPE` wherever LITERAL is a bare
hex/decimal integer literal and TYPE textually matches the enclosing
static's declared type - 445 occurrences across 25 files. Also
unwraps the now-redundant parens left behind on negated literals
(`-(1 as i32)` -> `-(1)` -> `-1`, 20 occurrences).

Same exclusions as the const phase: char-literal casts (`'x' as i32`)
are untouched since a char literal's type is always `char`, never
inferred from context; casts on anything other than a bare literal
(another const/static's name, a function call, a differently-typed
sub-expression) are left alone. Array-typed statics are unaffected by
construction - the declared type text (e.g. `[i32; 200]`) never
matches a per-element `LITERAL as i32` cast, so those go untouched.

Verified: cargo build (lib + doomgeneric_xlib bin) shows 0 new
errors; full warning-text diff against the current 103-warning
baseline is byte-identical. 3 consecutive Xvfb runs against
doom1.wad (-warp 1 1 -skill 3), no crash, screenshot-verified
rendering.

+++ rust-doomgeneric/src/i_joystick.rs
@@ -3,14 +3,14 @@ use crate::src::m_misc::M_snprintf;
 use crate::src::stdint_types::size_t;
 
 pub const NUM_VIRTUAL_BUTTONS: i32 = 10;
-static mut usejoystick: i32 = 0 as i32;
-static mut joystick_index: i32 = -(1 as i32);
-static mut joystick_x_axis: i32 = 0 as i32;
-static mut joystick_x_invert: i32 = 0 as i32;
-static mut joystick_y_axis: i32 = 1 as i32;
-static mut joystick_y_invert: i32 = 0 as i32;
-static mut joystick_strafe_axis: i32 = -(1 as i32);
-static mut joystick_strafe_invert: i32 = 0 as i32;
+static mut usejoystick: i32 = 0;
+static mut joystick_index: i32 = -1;
+static mut joystick_x_axis: i32 = 0;
+static mut joystick_x_invert: i32 = 0;
+static mut joystick_y_axis: i32 = 1;
+static mut joystick_y_invert: i32 = 0;
+static mut joystick_strafe_axis: i32 = -1;
+static mut joystick_strafe_invert: i32 = 0;
 static mut joystick_physical_buttons: [i32; 10] = [
     0 as i32,
     1 as i32,
